refactor: split theme script into modules
